Create, update, and delete tests moved out of DataTest
