+++ sdk/appcenter-data/src/test/java/com/microsoft/appcenter/data/DataReadTest.java
@@ -308,5 +308,4 @@ public class DataReadTest extends AbstractDataTest {
         verifyNoMoreInteractions(mHttpClient);
         assertEquals(createdDocument.getDeserializedValue(), document.getDeserializedValue());
     }
-
 }
